+++ libs/framework-core/source/ftrack_framework_core/event/__init__.py
@@ -320,6 +320,25 @@ class Publish(object):
         event_topic = constants.event.HOST_VERIFY_PLUGINS_TOPIC
         return self._publish_event(event_topic, data, callback)
 
+    def host_sync_tool_config(
+        self,
+        host_id,
+        tool_config,
+        callback=None,
+        mode=constants.event.LOCAL_EVENT_MODE,
+    ):
+        '''
+        Publish an event with topic
+        :const:`~ftrack_framework_core.constants.event.HOST_VERIFY_PLUGINS_TOPIC`
+        '''
+        data = {
+            'host_id': host_id,
+            'tool_config': tool_config,
+        }
+
+        event_topic = constants.event.HOST_SYNC_TOOL_CONFIG_TOPIC
+        return self._publish_event(event_topic, data, callback, mode)
+
 
 class Subscribe(object):
     '''Class with all the events subscribed by the framework'''
@@ -451,5 +470,15 @@ class Subscribe(object):
         '''
         event_topic = '{} and data.host_id={}'.format(
             constants.event.HOST_VERIFY_PLUGINS_TOPIC, host_id
+        )
+        return self._subscribe_event(event_topic, callback)
+
+    def host_sync_tool_config(self, host_id, callback=None):
+        '''
+        Subscribe to an event with topic
+        :const:`~ftrack_framework_core.constants.event.HOST_SYNC_TOOL_CONFIG_TOPIC`
+        '''
+        event_topic = '{} and data.host_id={}'.format(
+            constants.event.HOST_SYNC_TOOL_CONFIG_TOPIC, host_id
         )
         return self._subscribe_event(event_topic, callback)