+++ libs/qt/source/ftrack_qt/widgets/labels/__init__.py
@@ -1,0 +1,1 @@
+from ftrack_qt.widgets.labels.editable_label import EditableLabel